fix: 당첨 금액의 Long타입 적용으로 오버플로우 수정

+++ src/main/kotlin/lotto/domain/valueobject/PrizeMoney.kt
@@ -1,7 +1,7 @@
 package lotto.domain.valueobject
 
 data class PrizeMoney(
-    val money: Int,
+    val money: Long,
 ) {
     init {
         require(money >= 0) { ERROR_NEGATIVE_NUMBER }
